+++ Branch-SDK/src/main/java/io/branch/referral/Branch.java
@@ -67,7 +67,7 @@ import io.branch.referral.util.LinkProperties;
  * </pre>
  * -->
  */
-public class Branch implements BranchViewHandler.IBranchViewEvents, SystemObserver.AdsParamsFetchEvents, InstallListener.IInstallReferrerEvents {
+public class Branch implements BranchViewHandler.IBranchViewEvents, SystemObserver.AdsParamsFetchEvents, GooglePlayStoreAttribution.IInstallReferrerEvents {
     
     private static final String BRANCH_LIBRARY_VERSION = "io.branch.sdk.android:library:" + BuildConfig.VERSION_NAME;
     private static final String GOOGLE_VERSION_TAG = "!SDK-VERSION-STRING!" + ":" + BRANCH_LIBRARY_VERSION;
@@ -293,7 +293,7 @@ public class Branch implements BranchViewHandler.IBranchViewEvents, SystemObserv
     private static boolean bypassCurrentActivityIntentState_ = false;
 
     static boolean checkInstallReferrer_ = true;
-    private static long playStoreReferrerFetchTime = 1500;
+    private static long playStoreReferrerWaitTime = 1500;
     public static final long NO_PLAY_STORE_REFERRER_WAIT = 0;
     
     /**
@@ -512,7 +512,8 @@ public class Branch implements BranchViewHandler.IBranchViewEvents, SystemObserv
     }
     
     /**
-     * @deprecated This method is deprecated since play store referrer is enabled by default from v2.9.1.
+     * @deprecated This method is deprecated since INSTALL_REFERRER broadcasts were discontinued on 3/2020.
+     * And Branch SDK bundles Play Store Referrer library since v4.2.2
      * Please use {@link #setPlayStoreReferrerCheckTimeout(long)} instead.
      */
     public static void enablePlayStoreReferrer(long delay) {
@@ -520,17 +521,16 @@ public class Branch implements BranchViewHandler.IBranchViewEvents, SystemObserv
     }
     
     /**
-     * Since play store referrer broadcast from google play is few millisecond delayed Branch will delay the collecting deep link data on app install by {@link #playStoreReferrerFetchTime} millisecond
-     * This will allow branch to provide for more accurate tracking and attribution. This will delay branch init only the first time user open the app.
-     * This method allows to override the maximum wait time for play store referrer to arrive. Set it to {@link Branch#NO_PLAY_STORE_REFERRER_WAIT} if you don't want to wait for play store referrer
+     * Set timeout for Play Store Referrer library. Play Store Referrer library allows Branch to provide
+     * more accurate tracking and attribution. This delays Branch initialization only the first time user opens the app.
+     * This method allows to override the maximum wait time for play store referrer to arrive.
      * <p>
-     * Note:  as of our testing 4/2017  a 1500 milli sec wait time is enough to capture more than 90% of the install referrer case
      *
      * @param delay {@link Long} Maximum wait time for install referrer broadcast in milli seconds. Set to {@link Branch#NO_PLAY_STORE_REFERRER_WAIT} if you don't want to wait for play store referrer
      */
     public static void setPlayStoreReferrerCheckTimeout(long delay) {
         checkInstallReferrer_ = delay > 0;
-        playStoreReferrerFetchTime = delay;
+        playStoreReferrerWaitTime = delay;
     }
     
     /**
@@ -778,7 +778,7 @@ public class Branch implements BranchViewHandler.IBranchViewEvents, SystemObserv
         // BranchStrongMatchHelper.shutDown();
         // BranchViewHandler.shutDown();
         // DeepLinkRoutingValidator.shutDown();
-        // InstallListener.shutDown();
+        // GooglePlayStoreAttribution.shutDown();
         // InstantAppUtil.shutDown();
         // IntegrationValidator.shutDown();
         // ShareLinkManager.shutDown();
@@ -2432,7 +2432,7 @@ public class Branch implements BranchViewHandler.IBranchViewEvents, SystemObserv
         }
     }
 
-    private void initializeSession(final BranchReferralInitListener callback, boolean isReInitializing) {
+    private void initializeSession(final BranchReferralInitListener callback) {
         if ((prefHelper_.getBranchKey() == null || prefHelper_.getBranchKey().equalsIgnoreCase(PrefHelper.NO_STRING_VALUE))) {
             setInitState(SESSION_STATE.UNINITIALISED);
             //Report Key error on callback
@@ -2446,7 +2446,7 @@ public class Branch implements BranchViewHandler.IBranchViewEvents, SystemObserv
         }
 
         ServerRequestInitSession initRequest = getInstallOrOpenRequest(callback);
-        if (!isReInitializing && (getSessionReferredLink() == null || enableFacebookAppLinkCheck_)) {
+        if (initState_ == SESSION_STATE.UNINITIALISED && (getSessionReferredLink() == null || enableFacebookAppLinkCheck_)) {
             // Check if opened by facebook with deferred install data
             boolean appLinkRqSucceeded = DeferredAppLinkDataHandler.fetchDeferredAppLinkData(
                     context_, new DeferredAppLinkDataHandler.AppLinkFetchEvents() {
@@ -2482,12 +2482,11 @@ public class Branch implements BranchViewHandler.IBranchViewEvents, SystemObserv
 
         // !isFirstInitialization condition equals true only when user calls reInitSession()
 
-        if (getInitState() == SESSION_STATE.UNINITIALISED || isReInitializing || forceBranchSession) {
+        if (getInitState() == SESSION_STATE.UNINITIALISED || forceBranchSession) {
             registerAppInit(initRequest, false);
         } else if (callback != null) {
             // Else, let the user know session initialization failed because it's already initialized.
-            callback.onInitFinished(null,
-                    new BranchError("Session initialization failed.", BranchError.ERR_BRANCH_ALREADY_INITIALIZED));
+            callback.onInitFinished(null, new BranchError("Warning.", BranchError.ERR_BRANCH_ALREADY_INITIALIZED));
         }
     }
     
@@ -2504,9 +2503,19 @@ public class Branch implements BranchViewHandler.IBranchViewEvents, SystemObserv
             if (intentState_ != INTENT_STATE.READY  && isWaitingForIntent()) {
                 request.addProcessWaitLock(ServerRequest.PROCESS_WAIT_LOCK.INTENT_PENDING_WAIT_LOCK);
             }
-            if (checkInstallReferrer_ && request instanceof ServerRequestRegisterInstall && !InstallListener.unreportedReferrerAvailable) {
+
+            // Google Play Referrer lib should only be used once, so we use GooglePlayStoreAttribution.hasBeenUsed flag
+            // just in case user accidentally queues up a couple install requests at the same time. During later sessions
+            // request instanceof ServerRequestRegisterInstall = false
+            if (checkInstallReferrer_ && request instanceof ServerRequestRegisterInstall && !GooglePlayStoreAttribution.hasBeenUsed) {
                 request.addProcessWaitLock(ServerRequest.PROCESS_WAIT_LOCK.INSTALL_REFERRER_FETCH_WAIT_LOCK);
-                new InstallListener().captureInstallReferrer(context_, playStoreReferrerFetchTime, this);
+                new GooglePlayStoreAttribution().captureInstallReferrer(context_, playStoreReferrerWaitTime, this);
+
+                // GooglePlayStoreAttribution error are thrown synchronously, so we remove
+                // INSTALL_REFERRER_FETCH_WAIT_LOCK manually (see GooglePlayStoreAttribution.erroredOut)
+                if (GooglePlayStoreAttribution.erroredOut) {
+                    request.removeProcessWaitLock(ServerRequest.PROCESS_WAIT_LOCK.INSTALL_REFERRER_FETCH_WAIT_LOCK);
+                }
             }
         }
 
@@ -2517,6 +2526,8 @@ public class Branch implements BranchViewHandler.IBranchViewEvents, SystemObserv
         if (!requestQueue_.containsInitRequest()) {
             insertRequestAtFront(request);
             processNextQueueItem();
+        } else {
+            PrefHelper.Debug("Warning! Attempted to queue multiple init session requests");
         }
     }
     
@@ -2527,7 +2538,7 @@ public class Branch implements BranchViewHandler.IBranchViewEvents, SystemObserv
             request = new ServerRequestRegisterOpen(context_, callback);
         } else {
             // If no user this is an Install
-            request = new ServerRequestRegisterInstall(context_, callback, InstallListener.getInstallationID());
+            request = new ServerRequestRegisterInstall(context_, callback);
         }
         return request;
     }
@@ -2634,7 +2645,7 @@ public class Branch implements BranchViewHandler.IBranchViewEvents, SystemObserv
             PrefHelper.Debug(new BranchError("", BranchError.ERR_API_LVL_14_NEEDED).getMessage());
         }
     }
-    
+
     /*
      * Check for forced session restart. The Branch session is restarted if the incoming intent has branch_force_new_session set to true.
      * This is for supporting opening a deep link path while app is already running in the foreground. Such as clicking push notification while app in foreground.
@@ -2673,7 +2684,7 @@ public class Branch implements BranchViewHandler.IBranchViewEvents, SystemObserv
      * @see BranchError
      */
     public interface BranchReferralInitListener {
-        void onInitFinished(JSONObject referringParams, BranchError error);
+        void onInitFinished(@Nullable JSONObject referringParams, @Nullable BranchError error);
     }
     
     /**
@@ -2688,7 +2699,7 @@ public class Branch implements BranchViewHandler.IBranchViewEvents, SystemObserv
      * @see BranchError
      */
     public interface BranchUniversalReferralInitListener {
-        void onInitFinished(BranchUniversalObject branchUniversalObject, LinkProperties linkProperties, BranchError error);
+        void onInitFinished(@Nullable BranchUniversalObject branchUniversalObject, @Nullable LinkProperties linkProperties, @Nullable BranchError error);
     }
     
     
@@ -2702,7 +2713,7 @@ public class Branch implements BranchViewHandler.IBranchViewEvents, SystemObserv
      * @see BranchError
      */
     public interface BranchReferralStateChangedListener {
-        void onStateChanged(boolean changed, BranchError error);
+        void onStateChanged(boolean changed, @Nullable BranchError error);
     }
     
     /**
@@ -3655,7 +3666,8 @@ public class Branch implements BranchViewHandler.IBranchViewEvents, SystemObserv
                 branch.checkForAutoDeepLinkConfiguration();
             }
 
-            branch.initializeSession(callback, isReInitializing);
+            branch.initializeSession(callback);
+//            branch.initializeSession(callback, isReInitializing);
         }
 
         /**